M5: Fixed rendering issue by switching to tilemap texture baking instead of camera culling

+++ engine/renderer/Camera.h
@@ -6,7 +6,7 @@ struct Camera {
     float width = 0.0f;
     float height = 0.0f;
 
-    void Follow(float targetX, float targetY, float dt, float smoothing) {
+    void Follow(float targetX, float targetY, double dt, float smoothing) {
         float targetCameraX = targetX - width / 2.0f;
         float targetCameraY = targetY - height / 2.0f;
